implement very unfinished teleporters + fix mystinvector2 bug
teleporters are still missing:
in editor validation when placing
revalidation when moved
patches/code for handling teleporters being missing
icons

+++ PlusStudioLevelFormat/Datatypes.cs
@@ -68,8 +68,8 @@ namespace PlusStudioLevelFormat
 
         public MystIntVector2(int x, int y)
         {
-            _x = (byte)x;
-            _z = (byte)y;
+            _x = x;
+            _z = y;
         }
 
         public static MystIntVector2 one => new MystIntVector2(1, 1);
